Cleanup ServerHandler class with the goal of fixing all obvious memory leaks.
There are a few methods in here that are very difficult to understand.  Without
documentation on the protocol, it is hard to follow, so I'm not touching it yet... ;-)
Ultimately, this class will also need to be broken into a couple
different pieces. However, at this point, I'd rather focus on writing some
tests for what is already here.