Cleanup initialization methods in ServerHandler & add methods to create autoreleased instances of classes.
At this point, I'm not sure if the ServerHandler class is meant to be
a client, connecting to a remote repository, or a server to respond to
remote requests (or both?).

Finally changed 'initFrom' methods to 'initWith', since the object being
referred to will be used to create the instance, and will be retained.