Reorganized dir structure for use with/without Xcode, and to facilitate testing.
I like the way that the [touch-code][] projects are organized:
All of the source code goes in a separate dir, which is separate from the .xcodeproj
that references the code. This makes it really easy to build a framework
for testing on OS X, or to simply copy the files to a new Xcode project
(via drag-and-drop) for including in an iPhone project.

[touch-code]: http://code.google.com/p/touchcode/